requirements.txt
Flask
python-dotenv
pytz
gunicorn
# Optional database drivers:
# For MySQL/MariaDB support, uncomment the line below:
# mysql-connector-python
# SQLite is built-in with Python
